# api/openapi.yaml
openapi: 3.1.0
info:
  title: Lantai API
  version: 1.0.0-draft
  summary: 兰台 M2 REST 与传输契约
  description: |
    兰台只有一套 REST 接口。HTTP 路由、请求和视图由本文件定义，共享数据、清单与错误引用 schemas/ 权威定义。scripts/gen/openapi 打包为 api/gen/openapi.bundle.json；内部 handler 使用相同契约。

    /api/v1 仅加不改；持久业务创建、提交、著录和敏感命令执行要求 Idempotency-Key。会话/挑战/短时授权签发以及绑定既有上传的续传、完成、取消依各路由语义重试。著录修改要求带 If-Match。所有服务端时间为固定 UTC 毫秒。默认 JSON 上限 8 MiB，超限返回 SCHEMA_INVALID（request_body_too_large）；错误码的 HTTP 状态由统一注册表决定。
  license:
    name: GPL-3.0-only
    identifier: GPL-3.0-only
servers:
  - url: /
    description: 经网关同源访问；接口面前缀 /api/，传输面前缀 /xfer/（地址由服务端给出）
security:
  - bearerSession: []
  - cookieSession: []
...
paths:
  /api/v1/operations/{operation_id}:
    get:
      tags: [operations]
      operationId: getOperation
      summary: 查询持久操作状态
      description: |
        返回操作阶段、结果引用、是否可重试与下一步建议。结果引用按调用者
        当前读取权限过滤；无权知道的操作返回 404，不泄露存在性。
      parameters:
        - $ref: '#/components/parameters/OperationID'
        - $ref: '#/components/parameters/RequestID'
      responses:
        '200':
          description: 操作状态
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '../schemas/common/v1/operation.schema.json'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
  /api/v1/assets:
    get:
      operationId: searchAssets
      parameters:
        - in: query
          name: view
          required: false
          schema:
            default: brief
            enum:
              - brief
              - full
            type: string
        - in: query
          name: limit
          required: false
          schema:
            default: 50
            maximum: 100
            minimum: 1
            type: integer
        - in: query
          name: cursor
          required: false
          schema:
            type: string
        - in: query
          name: q
          required: false
          schema:
            type: string
        - in: query
          name: project_id
          required: false
          schema:
            $ref: ../schemas/common/v1/defs.schema.json#/$defs/ulid
        - in: query
          name: asset_type
          required: false
          schema:
            $ref: ../schemas/catalog/v1/manifest.schema.json#/$defs/asset_type
        - $ref: '#/components/parameters/RequestID'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryPage'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: searchAssets
  /api/v1/assets/{asset_id}:
    get:
      operationId: getAsset
      parameters:
        - in: query
          name: view
          required: false
          schema:
            default: brief
            enum:
              - brief
              - full
            type: string
        - $ref: '#/components/parameters/RequestID'
        - in: path
          name: asset_id
          required: true
          schema:
            $ref: ../schemas/common/v1/defs.schema.json#/$defs/ulid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetView'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: getAsset
  /api/v1/assets/{asset_id}/metadata:
    patch:
      description: If-Match is required and must be a quoted nonnegative revision, e.g. "0". Success returns the new revision in ETag.
      operationId: patchAssetMetadata
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/RequestID'
        - in: path
          name: asset_id
          required: true
          schema:
            $ref: ../schemas/common/v1/defs.schema.json#/$defs/ulid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetPatch'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetDescriptionView'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: patchAssetMetadata
  /api/v1/assets/{asset_id}/versions/{version_id}:
    get:
      description: Authority-backed exact read. Full includes the frozen manifest; brief omits it.
      operationId: getVersion
      parameters:
        - in: query
          name: view
          required: false
          schema:
            default: brief
            enum:
              - brief
              - full
            type: string
        - $ref: '#/components/parameters/RequestID'
        - in: path
          name: asset_id
          required: true
          schema:
            $ref: ../schemas/common/v1/defs.schema.json#/$defs/ulid
        - in: path
          name: version_id
          required: true
          schema:
            $ref: ../schemas/common/v1/defs.schema.json#/$defs/ulid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExactVersionView'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: getVersion
  /api/v1/assets/{asset_id}/versions/{version_id}/read-grants:
    post:
      description: Creates a fresh short-lived session-bound capability. It is not a persisted business command and does not require Idempotency-Key. Follow returned url with the same session; every new GET/Range is authorized again.
      operationId: issueReadGrant
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - in: path
          name: asset_id
          required: true
          schema:
            $ref: ../schemas/common/v1/defs.schema.json#/$defs/ulid
        - in: path
          name: version_id
          required: true
          schema:
            $ref: ../schemas/common/v1/defs.schema.json#/$defs/ulid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadGrantRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadGrant'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: issueReadGrant
...
  /api/v1/meta:
    get:
      operationId: getMeta
      parameters:
        - $ref: '#/components/parameters/RequestID'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meta'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      security: []
      summary: getMeta
  /api/v1/projects:
    get:
      operationId: listProjects
      parameters:
        - in: query
          name: view
          required: false
          schema:
            default: brief
            enum:
              - brief
              - full
            type: string
        - in: query
          name: limit
          required: false
          schema:
            default: 50
            maximum: 100
            minimum: 1
            type: integer
        - in: query
          name: cursor
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/RequestID'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPage'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: listProjects
    post:
      operationId: createProject
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectView'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: createProject
  /api/v1/projects/{key}:
    get:
      operationId: getProject
      parameters:
        - in: query
          name: view
          required: false
          schema:
            default: brief
            enum:
              - brief
              - full
            type: string
        - $ref: '#/components/parameters/RequestID'
        - in: path
          name: key
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectView'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: getProject
  /api/v1/sessions/current:
    delete:
      operationId: endCurrentSession
      parameters:
        - $ref: '#/components/parameters/RequestID'
      responses:
        "204":
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: endCurrentSession
  /api/v1/sessions/exchange:
    post:
      operationId: exchangeSession
      parameters:
        - $ref: '#/components/parameters/RequestID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      security: []
      summary: exchangeSession
  /api/v1/sessions/login:
    post:
      description: Human password plus current authenticator code. Browser channel requires an exact configured Origin and refuses cross-site Fetch metadata. Source rate limit uses the direct peer, never an untrusted forwarding header.
      operationId: loginSession
      parameters:
        - $ref: '#/components/parameters/RequestID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      security: []
      summary: loginSession
...
  /api/v1/types:
    get:
      operationId: listTypes
      parameters:
        - $ref: '#/components/parameters/RequestID'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypePage'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: listTypes
  /api/v1/types/{type}:
    get:
      operationId: getAssetType
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - in: path
          name: type
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeDetail'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: getAssetType
  /api/v1/uploads:
    post:
      description: Persistently fixes operation_id. Use each file part_url_template, substituting only {part_number}; do not construct transfer routes.
      operationId: createUpload
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUploadRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadView'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: createUpload
  /api/v1/uploads/{upload_id}:
    delete:
      description: Cancel the existing upload/commit operation through catalog. The existing upload identity is the retry key; no new business intent is created.
      operationId: cancelUpload
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - in: path
          name: upload_id
          required: true
          schema:
            $ref: ../schemas/common/v1/defs.schema.json#/$defs/ulid
      responses:
        "204":
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: cancelUpload
    get:
      operationId: getUpload
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - in: path
          name: upload_id
          required: true
          schema:
            $ref: ../schemas/common/v1/defs.schema.json#/$defs/ulid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadView'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: getUpload
  /api/v1/uploads/{upload_id}/check:
    post:
      operationId: checkUploadBlobs
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - in: path
          name: upload_id
          required: true
          schema:
            $ref: ../schemas/common/v1/defs.schema.json#/$defs/ulid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckBlobsRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlobStatusPage'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: checkUploadBlobs
  /api/v1/uploads/{upload_id}/commit:
    post:
      description: Committed is visible. Description_pending means only the first description remains to be completed; retry with the same key. No synthetic success or 202 is created by transport.
      operationId: commitUpload
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - in: path
          name: upload_id
          required: true
          schema:
            $ref: ../schemas/common/v1/defs.schema.json#/$defs/ulid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitResult'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: commitUpload
  /api/v1/uploads/{upload_id}/files/{sha256}/complete:
    post:
      operationId: completeUploadFile
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - in: path
          name: upload_id
          required: true
          schema:
            $ref: ../schemas/common/v1/defs.schema.json#/$defs/ulid
        - in: path
          name: sha256
          required: true
          schema:
            $ref: ../schemas/common/v1/defs.schema.json#/$defs/sha256_hex
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmptyRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlobGrant'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: completeUploadFile
  /api/v1/whoami:
    get:
      operationId: whoAmI
      parameters:
        - $ref: '#/components/parameters/RequestID'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityView'
          description: 成功
        default:
          $ref: '#/components/responses/Error'
      summary: whoAmI
...
components:
  securitySchemes:
    bearerSession:
      type: http
      scheme: bearer
      description: |
        会话令牌 lts_…（由长期凭据换取或由人登录得到，限定范围与有效期）。长期凭据 ltk_…
        只能换会话，不能直接调用业务接口；浏览器会话令牌只能经 Cookie 使用。Bearer 请求
        不做 CSRF 检查，但仍须通过正常授权。见 docs/contracts/identity.md。
    cookieSession:
      type: apiKey
      in: cookie
      name: lantai_session
      description: |
        网页的 HttpOnly/Secure/SameSite=Strict 会话。GET/HEAD/OPTIONS 以外的请求必须带与
        允许列表精确相同的 Origin（缺失即拒绝）并在 X-CSRF-Token 头中带与该会话相符的 CSRF
        令牌；同一请求不能同时带 Bearer 与会话 Cookie。
  schemas:
...
    AssetView:
      additionalProperties: false
      properties:
        asset_id:
          $ref: ../schemas/common/v1/defs.schema.json#/$defs/ulid
        asset_type:
          type: string
        description:
          $ref: '#/components/schemas/AssetDescriptionView'
        latest:
          $ref: '#/components/schemas/VersionView'
        project_id:
          $ref: ../schemas/common/v1/defs.schema.json#/$defs/ulid
      required:
        - asset_id
        - project_id
        - asset_type
        - description
        - latest
      type: object
...
    EmptyRequest:
      additionalProperties: false
      properties: {}
      type: object
...
    QueryPage:
      additionalProperties: false
      description: Both views return the security-filtered catalog projection. Full does not expand files/uses; use exact reads for the frozen manifest.
      properties:
        items:
          items:
            $ref: '#/components/schemas/QueryItem'
          type: array
        next_cursor:
          type: string
        state:
          additionalProperties: false
          properties:
            generation:
              minimum: 0
              type: integer
            high_water:
              minimum: 0
              type: integer
            rebuild_start:
              minimum: 0
              type: integer
          required:
            - generation
            - high_water
            - rebuild_start
          type: object
      required:
        - items
        - state
      type: object
...
  parameters:
...
    OperationID:
      name: operation_id
      in: path
      required: true
      schema:
        $ref: '../schemas/common/v1/defs.schema.json#/$defs/ulid'
...
    RequestID:
      name: X-Request-Id
      in: header
      required: false
      description: 客户端请求编号，仅用于日志关联，不参与请求摘要。
      schema:
        type: string
        maxLength: 128
...
  headers:
    RequestID:
      description: 服务端请求编号。
      schema:
        type: string
...
  responses:
...
    Error:
      description: 结构化错误；错误码、HTTP 状态、retryable 与 recovery_action 以错误码注册表为准。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestID'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '../schemas/common/v1/error.schema.json'
...